Break up code into several classes
